@@ -1,20 +1,19 @@
 name: flustars
 description: Flutter common utils library. SpUtil, ScreenUtil, WidgetUtil, DirectoryUtil, ImageUtil. SharedPreferences Util. Screen info & Screen adaptation. get image size.
-version: 0.3.3
-author: Sky24n <863764940@qq.com>
+version: 2.0.0
 homepage: https://github.com/Sky24n/flustars
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   # Dart common utils library. https://github.com/Sky24n/common_utils
-  common_utils: '>=1.2.1 <3.0.0'
+  common_utils: '>=2.0.0 <3.0.0'
   # https://github.com/Sky24n/sp_util
-  sp_util: '>=1.0.0 <3.0.0'
+  sp_util: '>=2.0.0 <3.0.0'
   # https://github.com/tekartik/synchronized.dart
-  synchronized: '>=1.5.1 <4.0.0'
+  synchronized: '>=3.0.0 <5.0.0'
   # https://github.com/flutter/plugins/tree/master/packages/path_provider
-  path_provider: '>=1.2.0 <3.0.0'
+  path_provider: '>=2.0.0 <4.0.0'
