@@ -12,8 +12,8 @@ import 'package:flutter/widgets.dart';
 
 /// Image Util.
 class ImageUtil {
-  ImageStreamListener _listener;
-  ImageStream _imageStream;
+  late ImageStreamListener _listener;
+  late ImageStream _imageStream;
 
   /// get image width height，load error throw exception.（unit px）
   /// 获取图片宽高，加载错误会抛出异常.（单位 px）
@@ -21,11 +21,11 @@ class ImageUtil {
   /// url network
   /// local url , package
   Future<Rect> getImageWH({
-    Image image,
-    String url,
-    String localUrl,
-    String package,
-    ImageConfiguration configuration,
+    Image? image,
+    String? url,
+    String? localUrl,
+    String? package,
+    ImageConfiguration? configuration,
   }) {
     Completer<Rect> completer = Completer<Rect>();
     _listener = ImageStreamListener(
@@ -36,7 +36,7 @@ class ImageUtil {
               0, 0, info.image.width.toDouble(), info.image.height.toDouble()));
         }
       },
-      onError: (dynamic exception, StackTrace stackTrace) {
+      onError: (dynamic exception, StackTrace? stackTrace) {
         _imageStream.removeListener(_listener);
         if (!completer.isCompleted) {
           completer.completeError(exception, stackTrace);
@@ -49,11 +49,12 @@ class ImageUtil {
         (localUrl == null || localUrl.isEmpty)) {
       return Future.value(Rect.zero);
     }
-    Image img = image != null
-        ? image
-        : ((url != null && url.isNotEmpty)
-            ? Image.network(url)
-            : Image.asset(localUrl, package: package));
+    Image? img = image;
+    if (img == null) {
+      img = (url != null && url.isNotEmpty)
+          ? Image.network(url)
+          : Image.asset(localUrl!, package: package);
+    }
     _imageStream = img.image.resolve(configuration ?? ImageConfiguration());
     _imageStream.addListener(_listener);
     return completer.future;
