@@ -12,8 +12,8 @@ import 'package:flutter/widgets.dart';
 /// Widget Util.
 class WidgetUtil {
   bool _hasMeasured = false;
-  double _width;
-  double _height;
+  double _width = 0;
+  double _height = 0;
 
   /// Widget rendering listener.
   /// Widget渲染监听.
@@ -21,11 +21,11 @@ class WidgetUtil {
   /// isOnce: true,Continuous monitoring  false,Listen only once.
   /// onCallBack: Widget Rect CallBack.
   void asyncPrepare(
-      BuildContext context, bool isOnce, ValueChanged<Rect> onCallBack) {
+      BuildContext context, bool isOnce, ValueChanged<Rect>? onCallBack) {
     if (_hasMeasured) return;
-    WidgetsBinding.instance.addPostFrameCallback((Duration timeStamp) {
-      RenderBox box = context.findRenderObject();
-      if (box != null && box.semanticBounds != null) {
+    WidgetsBinding.instance?.addPostFrameCallback((Duration timeStamp) {
+      RenderBox? box = getRenderBox(context);
+      if (box != null) {
         if (isOnce) _hasMeasured = true;
         double width = box.semanticBounds.width;
         double height = box.semanticBounds.height;
@@ -39,27 +39,34 @@ class WidgetUtil {
   }
 
   /// Widget渲染监听.
-  void asyncPrepares(bool isOnce, ValueChanged<Rect> onCallBack) {
+  void asyncPrepares(bool isOnce, ValueChanged<Rect>? onCallBack) {
     if (_hasMeasured) return;
-    WidgetsBinding.instance.addPostFrameCallback((Duration timeStamp) {
+    WidgetsBinding.instance?.addPostFrameCallback((Duration timeStamp) {
       if (isOnce) _hasMeasured = true;
-      if (onCallBack != null) onCallBack(null);
+      if (onCallBack != null) onCallBack(Rect.zero);
     });
   }
 
   ///get Widget Bounds (width, height, left, top, right, bottom and so on).Widgets must be rendered completely.
   ///获取widget Rect
   static Rect getWidgetBounds(BuildContext context) {
-    RenderBox box = context.findRenderObject();
-    return (box != null && box.semanticBounds != null)
-        ? box.semanticBounds
-        : Rect.zero;
+    RenderBox? box = getRenderBox(context);
+    return box?.semanticBounds ?? Rect.zero;
+  }
+
+  static RenderBox? getRenderBox(BuildContext context) {
+    RenderObject? renderObject = context.findRenderObject();
+    RenderBox? box;
+    if (renderObject != null) {
+      box = renderObject as RenderBox;
+    }
+    return box;
   }
 
   ///Get the coordinates of the widget on the screen.Widgets must be rendered completely.
   ///获取widget在屏幕上的坐标,widget必须渲染完成
   static Offset getWidgetLocalToGlobal(BuildContext context) {
-    RenderBox box = context.findRenderObject();
+    RenderBox? box = getRenderBox(context);
     return box == null ? Offset.zero : box.localToGlobal(Offset.zero);
   }
 
@@ -72,11 +79,11 @@ class WidgetUtil {
   /// url network
   /// local url , package
   static Future<Rect> getImageWH({
-    Image image,
-    String url,
-    String localUrl,
-    String package,
-    ImageConfiguration configuration,
+    Image? image,
+    String? url,
+    String? localUrl,
+    String? package,
+    ImageConfiguration? configuration,
   }) {
     if (image == null &&
         (url == null || url.isEmpty) &&
@@ -84,11 +91,12 @@ class WidgetUtil {
       return Future.value(Rect.zero);
     }
     Completer<Rect> completer = Completer<Rect>();
-    Image img = image != null
-        ? image
-        : ((url != null && url.isNotEmpty)
-            ? Image.network(url)
-            : Image.asset(localUrl, package: package));
+    Image? img = image;
+    if (img == null) {
+      img = (url != null && url.isNotEmpty)
+          ? Image.network(url)
+          : Image.asset(localUrl!, package: package);
+    }
     img.image
         .resolve(configuration ?? ImageConfiguration())
         .addListener(ImageStreamListener(
@@ -98,7 +106,7 @@ class WidgetUtil {
                   info.image.width.toDouble(), info.image.height.toDouble()));
             }
           },
-          onError: (dynamic exception, StackTrace stackTrace) {
+          onError: (dynamic exception, StackTrace? stackTrace) {
             if (!completer.isCompleted) {
               completer.complete(Rect.zero);
             }
@@ -117,11 +125,11 @@ class WidgetUtil {
   /// local url (full path/全路径，example："assets/images/ali_connors.png"，""assets/images/3.0x/ali_connors.png"" );
   /// package
   static Future<Rect> getImageWHE({
-    Image image,
-    String url,
-    String localUrl,
-    String package,
-    ImageConfiguration configuration,
+    Image? image,
+    String? url,
+    String? localUrl,
+    String? package,
+    ImageConfiguration? configuration,
   }) {
     if (image == null &&
         (url == null || url.isEmpty) &&
@@ -129,11 +137,12 @@ class WidgetUtil {
       return Future.error("image is null.");
     }
     Completer<Rect> completer = Completer<Rect>();
-    Image img = image != null
-        ? image
-        : ((url != null && url.isNotEmpty)
-            ? Image.network(url)
-            : Image.asset(localUrl, package: package));
+    Image? img = image;
+    if (img == null) {
+      img = (url != null && url.isNotEmpty)
+          ? Image.network(url)
+          : Image.asset(localUrl!, package: package);
+    }
     img.image
         .resolve(configuration ?? ImageConfiguration())
         .addListener(ImageStreamListener(
@@ -143,7 +152,7 @@ class WidgetUtil {
                   info.image.width.toDouble(), info.image.height.toDouble()));
             }
           },
-          onError: (dynamic exception, StackTrace stackTrace) {
+          onError: (dynamic exception, StackTrace? stackTrace) {
             if (!completer.isCompleted) {
               completer.completeError(exception, stackTrace);
             }
